separate library indexing from main indexing (bad)
