+++ enigma/src/main/java/org/quiltmc/enigma/api/analysis/index/jar/JarIndexer.java
@@ -8,6 +8,7 @@ import org.quiltmc.enigma.api.translation.representation.entry.FieldDefEntry;
 import org.quiltmc.enigma.api.translation.representation.entry.FieldEntry;
 import org.quiltmc.enigma.api.translation.representation.entry.MethodDefEntry;
 import org.quiltmc.enigma.api.translation.representation.entry.MethodEntry;
+import org.quiltmc.enigma.impl.analysis.index.AbstractJarIndex;
 
 public interface JarIndexer {
 	default void indexClass(ClassDefEntry classEntry) {
