+++ enigma/src/main/java/org/quiltmc/enigma/api/analysis/index/jar/MainJarIndex.java
@@ -1,0 +1,29 @@
+package org.quiltmc.enigma.api.analysis.index.jar;
+
+import org.quiltmc.enigma.impl.analysis.index.AbstractJarIndex;
+
+public class MainJarIndex extends AbstractJarIndex {
+	public MainJarIndex(JarIndexer... indexers) {
+		super(indexers);
+	}
+
+	/**
+	 * Creates an empty index, configured to use all built-in indexers.
+	 * @return the newly created index
+	 */
+	public static JarIndex empty() {
+		EntryIndex entryIndex = new EntryIndex();
+		InheritanceIndex inheritanceIndex = new InheritanceIndex(entryIndex);
+		ReferenceIndex referenceIndex = new ReferenceIndex();
+		BridgeMethodIndex bridgeMethodIndex = new BridgeMethodIndex(entryIndex, inheritanceIndex, referenceIndex);
+		PackageVisibilityIndex packageVisibilityIndex = new PackageVisibilityIndex();
+		EnclosingMethodIndex enclosingMethodIndex = new EnclosingMethodIndex();
+		LambdaIndex lambdaIndex = new LambdaIndex();
+		return new MainJarIndex(entryIndex, inheritanceIndex, referenceIndex, bridgeMethodIndex, packageVisibilityIndex, enclosingMethodIndex, lambdaIndex);
+	}
+
+	@Override
+	public String getTranslationKey() {
+		return "progress.jar.indexing.jar";
+	}
+}
