+++ enigma/src/main/java/org/quiltmc/enigma/api/translation/mapping/IndexEntryResolver.java
@@ -28,6 +28,7 @@ public class IndexEntryResolver implements EntryResolver {
 
 	private final IndexTreeBuilder treeBuilder;
 
+	// todo stupid
 	public IndexEntryResolver(JarIndex index) {
 		this.entryIndex = index.getIndex(EntryIndex.class);
 		this.inheritanceIndex = index.getIndex(InheritanceIndex.class);
